Switch from no parity to space parity for TFMini
The datasheet for the TFMini LIDAR lists the serial parity setting as "0" which I read as meaning no parity but, given other hints elsewhere in the datasheet, I now think this is actually space parity which pulls the parity bit of the character low in all cases. I've modified the code to use this setting instead if kNone.